[WIP] feat: implement App component with reactive state and input handling
