+++ examples/playground/src/main.ts
@@ -1,50 +1,6 @@
-import { createApp, reactive } from "chibivue"
+import { createApp } from "chibivue";
+import App from "./App.vue"
 
-const app = createApp({
-  setup() {
-    const state = reactive({ message: "Hello chibivue!", input: "" })
-    const changeMessage = () => {
-      state.message += "!"
-    }
-
-    const handleInput = (e: InputEvent) => {
-      state.input = (e.target as HTMLInputElement)?.value ?? ""
-    }
-
-    return { state, changeMessage, handleInput }
-  },
-
-  template: `
-    <div class="container" style="text-align: center">
-      <h2>{{ state.message }}</h2>
-      <img
-        width="150px"
-        src="https://upload.wikimedia.org/wikipedia/commons/thumb/9/95/Vue.js_Logo_2.svg/1200px-Vue.js_Logo_2.svg.png"
-        alt="Vue.js Logo"
-      />
-      <p><b>chibivue</b> is the minimal Vue.js</p>
-
-      <button @click="changeMessage">Click me!</button>
-
-      <br />
-
-      <label>
-        Input Data
-        <input @input="handleInput" />
-      </label>
-
-      <p>input value: {{ state.input }}</p>
-
-      <style>
-        .container {
-          height: 100vh;
-          padding: 16px;
-          background-color: #becdbe;
-          color: #2c3e50;
-        }
-      </style>
-    </div>
-  `,
-})
+const app = createApp(App)
 
 app.mount("#app")
