+++ examples/playground/vite.config.js
@@ -1,12 +1,16 @@
-import path from 'node:path'
-import { fileURLToPath } from 'node:url'
-import { defineConfig } from 'vite'
+import path from "node:path"
+import { fileURLToPath } from "node:url"
+import { defineConfig } from "vite"
+
+import chibivue from "../../packages/@extensions/vite-plugin-chibivue"
 
 const dirname = path.dirname(fileURLToPath(new URL(import.meta.url)))
+
 export default defineConfig({
   resolve: {
     alias: {
-      chibivue: path.resolve(dirname, '../../packages'),
+      chibivue: path.resolve(dirname, "../../packages")
     },
   },
+  plugins: [chibivue()],
 })
