+++ packages/@extensions/vite-plugin-chibivue/index.ts
@@ -1,0 +1,11 @@
+import type { Plugin } from "vite"
+
+export default function vitePluginChibivue(): Plugin {
+  return {
+    name: "vite:chibivue",
+
+    transform(code, id) {
+      return { code }
+    },
+  }
+}
